+++ src/senaite/core/permissions/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+#
+# flake8: noqa
+
+from .worksheet.permissions import AddWorksheet
+from .worksheet.permissions import AddWorksheetTemplate
+from .worksheet.permissions import EditWorksheet
+from .worksheet.permissions import ManageWorksheets
+from .worksheet.permissions import TransitionRejectWorksheet
+from .worksheet.permissions import TransitionRemoveWorksheet
+from .worksheet.permissions import WorksheetAddAttachment